TightBindingToolbox.BdGModel

TightBindingToolbox.BdG.BdGModel — Type

`BdGModel` is a data type representing a general Tight Binding system with pairing.

# Attributes
- `uc_hop  ::  UnitCell`: the Unit cell of the lattice with the hoppings.
- `uc_pair ::  UnitCell`: the Unit cell of the lattice with the pairings.
- `bz      ::  BZ`: The discretized Brillouin Zone.
- `Ham     ::  Hamiltonian`: the Hamiltonian at all momentum-points.
- `T       ::  Float64`: the temperature of the system.
- `filling ::  Float64`: The filling of the system.
- `mu      ::  Float64`: The chemical potential of the system.
- `stat    ::  Int64` : ±1 for bosons and fermions.
- `gap     ::  Float64` : the energy gap of excitations at the given filling.
- `Gk      ::  Array{Matrix{ComplexF64}}` : An Array (corresponding to the grid of k-points in `BZ`) of Greens functions.
- `Fk      ::  Array{Matrix{ComplexF64}}` : An Array (corresponding to the grid of k-points in `BZ`) of anomalous Greens functions.

Initialize this structure using
```julia
BdGModel(uc_hop::UnitCell, uc_pair::UnitCell, bz::BZ, Ham::Hamiltonian ; T::Float64=1e-3, filling::Float64=-1.0, mu::Float64=0.0, stat::Int64=-1)
```
You can either input a filling, or a chemical potential. The corresponding μ for a given filling, or filling for a given μ is automatically calculated.

TightBindingToolbox.TBModel.FindFilling — Method

FindFilling(M::BdGModel) --> Float64
FindFilling(mu::Float64, M::BdGModel)

Find filling at given temperature and chemical potential that takes BdGModel object as argument instead of Hamiltonian, since for a BdG case, the filling depends on the wavefunctions also. Because of this, if you want to calculate the filling at a different chemical potential, you have to modify the Hamiltonian, the UnitCells, rediagonalize, and recalculate eveyrthing.

TightBindingToolbox.TBModel.FreeEnergy — Method

FreeEnergy(M::BdGModel; F0::Float64 = 0.0) --> Float64

Calculate the free energy of the given BdGModel.

TightBindingToolbox.TBModel.GetFilling! — Method

GetFilling!(M::BdGModel)

Function to get filling given a chemical potential!

TightBindingToolbox.TBModel.GetGap! — Method

GetGap!(M::BdGModel)

Calculate the BdG gap of the system.

TightBindingToolbox.TBModel.GetGk! — Method

GetGk!(M::BdGModel)

Finding the Greens functions, and anomalous greens functions in momentum space at some chemical potential.

TightBindingToolbox.TBModel.GetGr! — Method

GetGr!(M::BdGModel)

Finding the equal-time Greens functions and anomalous Greens function in real space of a BdGModel.

TightBindingToolbox.TBModel.GetMu! — Method

GetMu!(M::BdGModel ;  tol::Float64=0.001)

Function to get the correct chemical potential given a filling.

TightBindingToolbox.TBModel.GetRSEnergy — Method

GetRSEnergy(BdGModel) --> Float64

Returns the total energy of the BdG model including decomposed interactions. Computed from real space bonds.

TightBindingToolbox.TBModel.SolveModel! — Method

SolveModel!(M::BdGModel)

one-step function to find all the attributes in BdGModel after it has been initialized.