TightBindingToolbox.Lattice

TightBindingToolbox.LatticeStruct.Lattice — Type

Lattice{T} is a data type representing a general real-space lattice constructed out of a UnitCell{T}.

Attributes

uc :: UnitCell{T}: Unit Cell of the lattice.

size :: Vector{Int64}: size of the lattice in units of the UnitCell primitives.

length :: Int64: total number of sites.

sites :: Dict{ Tuple{Int64, Vector{Int64}}, Int64} : a dictionary with key = (sublattice, offset), and the corresponding value being the site number on the lattice.

positions :: Vector{ Vector{Float64}}: real-space positions of all the lattice sites.

fields :: Vector{ Vector{Float64}}: the fields on all the lattice sites.

bondSites :: Matrix{Int64}: a matrix with lattice.length rows such that bondSites[s, i] is the site number of the ith neighbour of site-s on the lattice.

bondDists :: Matrix{Float64}: a matrix with lattice.length rows such that bondDists[s, i] is the distance to the ith neighbour of site-s on the lattice.

bondLabels :: Matrix{String}: a matrix with lattice.length rows such that bondLabels[s, i] is the label of the bond to the ith neighbour of site-s on the lattice.

bondMats :: Matrix{Array{ComplexF64, T}}: a matrix with lattice.length rows such that bondMats[s, i] is the Array{ComplexF64, T} of the bond connecting to the ith neighbour of site-s on the lattice.

Initialize this structure using

Lattice( uc::UnitCell{T}, size::Vector{Int64} ; null_dist::Float64 = -1.0, null_label::String = "-")

where null_dist and null_label are used for bonds which are not allowed due to given boundary conditions, but still tracked in the code.

TightBindingToolbox.LatticeStruct.ApplyBCToSite — Method

ApplyBCToSite(site::Vector{Int64}, L::Vector{Int64}, BC::Vector{ComplexF64}) -->  Vector{Int64}

Returns the effective real-space offset of site given a lattice size L, and boundary conditions BC.

TightBindingToolbox.LatticeStruct.FillBonds! — Method

FillBonds!(lattice::Lattice{T} ; null_dist::Float64 = -1.0, null_label::String = "-" ) where {T}

Fills all the bond information in the lattice using the bonds in UnitCell.

TightBindingToolbox.LatticeStruct.FillLattice! — Method

FillLattice!( lattice::Lattice{T} ; null_dist::Float64 = -1.0, null_label::String = "-") where {T}

Wrapper function to fill both site and bond information in the lattice.

TightBindingToolbox.LatticeStruct.FillSites! — Method

FillSites!(lattice::Lattice{T})

Fills all the sites, positions, and fields of the lattice using information in the UnitCell.

TightBindingToolbox.LatticeStruct.GetBCPhase — Method

GetBCPhase(site::Vector{Int64}, L::Vector{Int64}, BC::Vector{ComplexF64}) --> ComplexF64

Returns the effective phase of site given a lattice size L, and boundary conditions BC.

TightBindingToolbox.LatticeStruct.GetMaxCoordinationNumber — Method

GetMaxCoordinationNumber(uc::UnitCell) --> Int64

Returns the maximum coordination number amongs all the sublattices of the UnitCell.

TightBindingToolbox.DesignLattice.CreateLattice — Method

CreateLattice(uc::UnitCell{T}, param::Param{T}, size::Vector{Int64} , index::Int64=length(param.value) ; null_dist::Float64 = -1.0, null_label::String = "-") --> Lattice{T} where {T}
CreateLattice(uc::UnitCell{T}, params::Vector{Param{T}}, size::Vector{Int64} , indices::Vector{Int64}=length.(getproperty.(params, :value)) ; null_dist::Float64 = -1.0, null_label::String = "-") :: Lattice{T} where {T}

Creates a lattice using Param objetcs given in params.

TightBindingToolbox.DesignLattice.ModifyLattice! — Method

ModifyLattice!(lattice::Lattice{T}, param::Param{T}) where {T}
ModifyLattice!(lattice::Lattice{T}, params::Vector{Param{T}}) where {T}

Modifies a lattice when the Param objects given in params are modified.

TightBindingToolbox.DesignLattice.RemoveLatticeBonds! — Method

RemoveLatticeBonds!(lattice::Lattice{T} , label::String ; null_dist::Float64 = -1.0, null_label::String = "-") where {T}

Removes a lattice bond with the given label.

TightBindingToolbox.DesignLattice.ScaleLattice! — Method

ScaleLattice!(lattice::Lattice{T}, param::Param{T}) where {T}
ScaleLattice!(lattice::Lattice{T}, params::Vector{Param{T}}) where {T}

Scales a lattice bond assuming that the Param objects got their strengths modified.

TightBindingToolbox.DesignLattice.ScaleLatticeBonds! — Method

ScaleLatticeBonds!(lattice::Lattice{T} , label::String , scaling::Float64) where {T}

Scales a lattice bond with the given label and by the given scaling.

Flux

TightBindingToolbox.Flux.CheckGaugeValidity — Method

CheckGaugeValidity(lat::Lattice{T}, A::Function ; _kwargs::Dict = Dict(), accuracy::Int64 = 5) --> Bool 

Checks if the gauge vector potential A is valid for the lattice lat under periodic boundary conditions.

TightBindingToolbox.Flux.GetBondPhase — Method

GetBondPhase(A::Function, r1::Vector{Float64}, r2::Vector{Float64} ; n::Int64 = 100, _kwargs::Dict = Dict()) --> ComplexF64

Returns the tight binding phase on a bond between two sites r1 and r2 on a lattice with a given gauge vector potential A. The phase is calculated by integrating the vector potential along the bond.

TightBindingToolbox.Flux.GetStringPhase — Method

GetStringPhase(Monopoles::Vector{Vector{Float64}}, r1::Vector{Float64}, r2::Vector{Float64} ;  flux::Float64 = Float64(pi)) --> ComplexF64

Get the Dirac string between the given Monopoles phase cutting the bond on the lattice between sites r1 and r2.

TightBindingToolbox.Flux.InsertMonopolePair! — Method

InsertMonopolePair!(lat::Lattice{T}, Monopoles::Vector{Vector{Float64}} ; flux::Float64 = Float64(pi) ) 

Inserts a pair of monopole <–> Anti-Monopole on the lattice lat at the given positions Monopoles with the given flux flux through a Dirac

TightBindingToolbox.Flux.ModifyGauge! — Method

ModifyGauge!(lat::Lattice{T}, A::Function ; n::Int64 = 100, _kwargs::Dict = Dict()) where{T}

Modifies the lattice hoppings using the fixed gauge vector potential A on the lattice lat by multiplying the bond matrices with the corresponding phase factors.