Utilities

Useful Functions

TightBindingToolbox.Useful.BinarySearch — Method

BinarySearch(target::Float64, xRange::Tuple{Float64, Float64}, f::T, args::Tuple ; tol::Float64=0.001)

General function implementing Binary search on a monotonic function f(x, args...)=target, in the range x ∈ xRange, with tolerance tol.

TightBindingToolbox.Useful.Central_Diff — Method

f : Can be an array of anything (even static arrays) delta : vector of real numbers with length = spatial dimensions = rank of f –-> the displacement vector when calculating the gradient PBC : vector of boolean with length = spatial dimensions = rank of f –-> if the ith dimension has PBC or not.

TightBindingToolbox.Useful.DeriDistFunction — Method

DeriDistFunction(E ; T::Float64, mu::Float64=0.0, stat::Int64=-1)

derivative of DistFunction w.r.t the energy. stat=1 –> Bose-Einstein distribution, and stat=-1 –> Fermi distribution.

TightBindingToolbox.Useful.DistFunction — Method

DistFunction(E ; T::Float64, mu::Float64=0.0, stat::Int64=-1)

Distribution function. stat=1 –> Bose-Einstein distribution, and stat=-1 –> Fermi distribution.

TightBindingToolbox.Useful.GetAllOffsets — Method

GetAllOffsets(OffsetRange::Int64, dim::Int64) --> Vector{Vector{Int64}}

Given a range, returns the set of all possible Bond offsets such that each element of the offset vector lies in [-OffsetRange, OffsetRange].

TightBindingToolbox.Useful.GetIndexPath — Method

GetIndexPath(start::Vector{Int64}, ending::Vector{Int64} ; exclusive::Bool=true) --> Vector{Vector{Int64}}

Returns a path in index-space of a discretized lattice which joins the two sets of indices start and ending. If the input exclusive is set to true, the returned path will NOT contain the ending point itself. Works in any dimensions.

TightBindingToolbox.Useful.Meshgrid — Method

Meshgrid(grid::Vector{Int64} ; starts::Vector{Int64} = zeros(Int64, length(grid)))

returns a meshgrid of (i1, i2, ..., in) such that ij ∈ [starts[j], starts[j] + 1, ..., grid[j]] ∀ j in [1, 2, ..., n = length(grid)]

TightBindingToolbox.Useful.SegmentIntersection — Method

SegmentIntersection(SegmentPoints::Vector{Vector{Float64}}, LinePoints::Vector{Vector{Float64}})-->Float64

Given two points on a line and two points on a segment, returns the parameter t such that the intersection point is given by (1-t) * SegmentPoints[1] + t * SegmentPoints[2]. If the intersection point is not on the segment, returns a value outside the range [0, 1].

TightBindingToolbox.Useful.VecAngle — Method

VecAngle(a::Vector{Float64}, b::Vector{Float64})

returns the angle b/w two given vectors.

Spin Matrices

TightBindingToolbox.SpinMatrices.HermitianBasis — Method

HermitianBasis(localDim::Int64) --> Vector{Matrix{ComplexF64}}

Returns a vector basis of Traceless, localDim x localDim, Hermitian matrices, normalized such that $Tr(T^{a †} . T^{b}) = (1/2)δ_{ab}$, along with the identity.

TightBindingToolbox.SpinMatrices.SpinMats — Method

SpinMats(spin::Rational{Int64}) --> Vector{Matrix{ComplexF64}}

Returns the 3 generators of SU(2) in the given spin-representation, along with the Identity matrix of d = 2*spin+1, in the format of [Sx, Sy, Sz, Id]