TightBindingToolbox.conduct

TightBindingToolbox.conduct.Conductivity — Type

Conductivity is a data type to track electrical conductivity of a tight-binding model.

Attributes

M :: Model: the tight-binding model whose conductivity is to be calculated.

omegas :: Vector{Float64}: The range of energies to integrate over to get the DC conductivity.

spread :: Float64: the disorder/spread in the real-frequency Greens function.

spectral:: Vector{Array{Matrix{ComplexF64}}}: the matrix Spectral function at the given frequencies, and all momentum points.

sigma :: Dict{String, Float64}: The dictionary containing the conductivity along the different directios.

Initialize this structure using

Conductivity(M::Model ; spread::Float64 = 1e-3)
Conductivity(M::Model , omegas::Vector{Float64} ; spread::Float64 = 1e-3)

You can either input a filling, or a chemical potential. The corresponding μ for a given filling, or filling for a given μ is automatically calculated.

TightBindingToolbox.conduct.GetConductivity! — Method

GetConductivity!(Cond::Conductivity, as::Vector{Int64}  ; get_velocity::Bool = false, get_spectral::Bool = false)

Calculates the full DC conductivity. Optional Booleans to calculate the velocity matrix of the Hamiltonian, calculate spectral functions also.

TightBindingToolbox.conduct.GetSpectralFunction! — Method

GetSpectralFunction!(cond::Conductivity)

Fills the spectral function in the conductivity class.

TightBindingToolbox.conduct.SpectralContribution — Method

SpectralContribution(v::Vector{Array{Matrix{ComplexF64}}}, H::Array{Matrix{ComplexF64}, N}, omega::Float64;  spread::Float64 = 1e-3, a::Int64 = 1):: ComplexF64 where {N}
SpectralContribution(v::Vector{Array{Matrix{ComplexF64}}}, A::Array{Matrix{ComplexF64}, N};  a::Int64 = 1):: ComplexF64 where {N}
SpectralContribution(v::Vector{Array{Matrix{ComplexF64}}}, H::Array{Matrix{ComplexF64}, N}, omega1::Float64, omega2::Float64;  spread::Float64 = 1e-3, a::Int64 = 1, b::Int64 = 2):: ComplexF64 where {N}
SpectralContribution(v::Vector{Array{Matrix{ComplexF64}}}, A1::Array{Matrix{ComplexF64}, N}, A2::Array{Matrix{ComplexF64}, N} ;  a::Int64 = 1, b::Int64 = 2):: ComplexF64 where {N}

The contribution to the conductivity coming from the spectral function for directions a (and b). Optionally can also pass the frequencies omega (and omega2).

TightBindingToolbox.conduct.SpectralFunction — Method

SpectralFunction(H::Array{Matrix{ComplexF64}, N}, omega::Float64 ;  spread::Float64 = 1e-3):: Array{Matrix{ComplexF64}} where {N}

Returns the matrix spectral function given the Hamiltonian, the frequency, and the spread.