TightBindingToolbox.susceptibility

TightBindingToolbox.suscep.Susceptibility — Type

Susceptibility is a data type representing the magnetic response, $χ^{ab}(Q , Ω)$ for a general tight-binding Model.

Attributes

M :: Model: the given model.

Qs :: Vector{Vector{Float64}}: the set of momentum points over which $χ^{ab}(Q , Ω)$ is calculated.

Omegas :: Vector{Float64}: the set of energies over which $χ^{ab}(Q , Ω)$ is calculated.

Spread :: Float64 : the finite spread when summing over delta functions.

chis :: Dict: a dictionary containing $χ^{ab}(Q , Ω)$ for the different directions e.g. chis["xx"] etc.

Initialize this structure using

Susceptibility(M::Model , Omegas::Vector{Float64} ;  eta::Float64 = 1e-2) = new{}(M, [], Omegas, eta, Dict())
Susceptibility(M::Model , Qs::Vector{Vector{Float64}}, Omegas::Vector{Float64} ;  eta::Float64 = 1e-2) = new{}(M, Qs, Omegas, eta, Dict())

TightBindingToolbox.suscep.FillChis! — Method

FillChis!(chi::Susceptibility; fill_BZ::Bool=false, a::Int64=3, b::Int64=3)

function to calculate susceptibility at a all given Ω=Omegas, but for all Q present in the given path, and along a fixed direction given by a and b.

TightBindingToolbox.suscep.FindReducedChi — Method

FindReducedChi(Q::Vector{Float64}, Omega::Float64 , M::Model; a::Int64=3, b::Int64=3, eta::Float64=1e-2) --> ComplexF64

function to calculate susceptibility at a fixed Ω=Omega, and Q, and along a fixed direction given by a and b.

TightBindingToolbox.suscep.nF_factor — Method

nF_factor(Q_index::Vector{Int64}, Omega::Float64 , M::Model; eta::Float64=1e-2) --> Matrix{Matrix{ComplexF64}}

function to calculate $(nF(E(k)) - nF(E(k+Q))) / (Ω + i * η - (E(k+Q) - E(k)))$ for a general multi-band system, at all k-points.